// src/app/models/ordre.ts
export interface Ordre {
  typeOrdre: string;
  numeroOrdre: string;
  codeAR: string;
  codeDocument: string;
  typeDocument: string;
  typeAbonnement: string;
  typeTransport: string;
  typeRemorque: string;
  codeSerie: string;
  cleVoyage: string;
  compteur: number;
  compteurDepart: number;
  compteurArrivee: number;
  codeTournee: string;
  numeroBordereau: string;
  codeVoyage: string;
  codeExpediteur: string;
  codeSociete: string;
  codeSite: string;
  codeTiers: string;
  tiersFacture: string;
  codeDestinataire: string;
  correspondantPrestataire: string;
  codePrestataire: string;
  dateSaisie: string;
  dateCreation: string;
  dateModification: string;
  dateValidation1: string;
  dateDepartPrevue: string;
  dateFinPrevue: string;
  semaineDepart: number;
  semaineFin: number;
  dateFacturation: string;
  dateTournee: string;
  dateEnvoi: string;
  datePlanification: string;
  dateBordereau: string;
  periodePlanification: string;
  dateMassification: string;
  dateEdition: string;
  paysDepart: string;
  paysArrivee: string;
  regionDepart: string;
  regionArrivee: string;
  villeDepartId: string;
  villeArriveeId: string;
  villeDepartNom: string;
  villeArriveeNom: string;
  codePostalDepart: string;
  codePostalArrivee: string;
  zoneDepart: string;
  zoneArrivee: string;
  zoneTarifaireDepart: string;
  zoneTarifaireArrivee: string;
  pointDepartIntermediaire: string;
  pointArriveeIntermediaire: string;
  kilometrage1: number;
  tempsEstimeTrajet: string;
  tempsAssocie: number;
  conditionsLivraison: string;
  conditionsPriseEnCharge: string;
  documentsAnnexes: string;
  plombDepart: string;
  plombArrivee: string;
  uniteTaxation: string;
  taxeManuelle: boolean;
  codeTVA: string;
  modeReglement: string;
  montantAcompte: number;
  facturationUnique: boolean;
  nonFacturable: boolean;
  numeroOrdreAvoir: string;
  memoFacturation: string;
  avoir: boolean;
  editionReception: boolean;
  timbreFiscal: boolean;
  importeExporte: boolean;
  aFacturer: boolean;
  cleTarifaire: string;
  dernierActionModification: string;
  memo: string;
  memoClient: string;
  memoTransporteur: string;
  memoAvoir: string;
  validationDocument: boolean;
  ordreValide: boolean;
  assigneA: string;
  massifie: boolean;
  massifiePar: string;
  indicateurInterne: boolean;
  utilisateurCreateur: string;
  utilisateurModificateur: string;
  ancienEtatOrdre: string;
  nombreLivraisons: number;
  etatOrdre: string;
  contrat: string;
  codearticle: string;
  selected: boolean;
  latitudeDepart:number;
  longitudeDepart:number;
  latitudeArrivee:number;
  longitudeArrivee:number;
    poid:number;
      nbrepalette:number;
  
}